create ObjectMapperConfig, delete JsonForman
