+++ src/main/java/faang/school/postservice/dto/comment/CommentEvent.java
@@ -1,6 +1,5 @@
 package faang.school.postservice.dto.comment;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -12,6 +11,5 @@ public class CommentEvent {
     private long postId;
     private long authorId;
     private long commentId;
-    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSSSS")
     private LocalDateTime timestamp;
 }
